Complete CustomerRst

+++ Order/src/main/java/com/helman/Webservice/CustomerRst.java
@@ -16,6 +16,8 @@ import com.helman.Dao.Customerdao;
 import com.helman.Entity.Customer;
 import com.helman.General.Logback;
 import javax.ws.rs.*;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.ArrayList;
@@ -25,43 +27,51 @@ import java.util.List;
 public class CustomerRst {
     private Customerdao customerdao = new Customerdao();
     private List<Customer> customerList = new ArrayList<>();
-    //http://localhost:8080/order/rest/customer/findall
+    Security sec = new Security();
+
+    //http://localhost:8080/order/rest/customer/all
     @GET
     @Path("/all")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response findall(){
+    public Response findall(@Context HttpHeaders headers){
         customerList.clear();
         customerList = customerdao.findall();
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.basicAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         try{
+            Customer customer = new Customer();
             FilterProvider filters = new SimpleFilterProvider().addFilter(
-                    "Customerfilter", SimpleBeanPropertyFilter.filterOutAllExcept("customerName", "contactLastName", "contactFirstName", "phone"));
+                    "Customerfilter", SimpleBeanPropertyFilter.filterOutAllExcept(customer.getfilters()));
             ObjectMapper mapper = new ObjectMapper();
             String customerJson = mapper.writer(filters).withDefaultPrettyPrinter().writeValueAsString(customerList);
-            //String customerJson = mapper(Listof obj customer);
+            //String customerJson = mapper(List of obj customer);
             Logback.logger.info("{}.{}|Try:Send all records to RESTful", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName());
             return Response.status(Response.Status.OK).entity(customerJson).build();
         }catch (JsonProcessingException e){
             Logback.logger.error("{}.{}|Exception:{}", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName(), e.getMessage());
             e.printStackTrace();
-            return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+            return Response.status(Response.Status.UNAUTHORIZED).entity(e.getMessage()).build();
         }
     }
     //http://localhost:8080/order/rest/customer/find/104
     @GET
     @Path("find/{customerNumber}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response findById(@PathParam("customerNumber") Integer customerNumber){
+    public Response findById(@PathParam("customerNumber") Integer customerNumber, @Context HttpHeaders headers){
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.basicAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         Customer customer = customerdao.findById(customerNumber);
         try{
-            FilterProvider filters = new SimpleFilterProvider().addFilter(
-              "Customerfilter", SimpleBeanPropertyFilter.filterOutAllExcept("customerNumber", "customerName", "contactLastName", "contactFirstName", "phone"));
+            FilterProvider filters = new SimpleFilterProvider().addFilter("Customerfilter", SimpleBeanPropertyFilter.filterOutAllExcept(customer.getfilters()));
             String customerJson = (new ObjectMapper()).writer(filters).withDefaultPrettyPrinter().writeValueAsString(customer);
             Logback.logger.info("{}.{}|Try: Send record to RESTful", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName());
             return Response.status(Response.Status.OK).entity(customerJson).build();
         }catch (JsonProcessingException e){
             Logback.logger.error("{}.{}|Exception:{}", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName(), e.getMessage());
             e.printStackTrace();
-            return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+            return Response.status(Response.Status.UNAUTHORIZED).entity(e.getMessage()).build();
         }
     }
     /*http://localhost:8080/order/rest/customer/insert
@@ -70,8 +80,8 @@ public class CustomerRst {
     "customerNumber": 102,
     "customerName": "Jack",
     "contactLastName": "maraghi",
-    "contactFirstName": "Shahrzad",
-    "phone": "+982188089786",
+    "contactFirstName": "Shahi",
+    "phone": "+982188089",
     "addressLine1":"Teh",
     "city": "Tehran",
     "country": "Iran"
@@ -80,7 +90,10 @@ public class CustomerRst {
     @Path("/insert")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response insert(Customer customer){
+    public Response insert(Customer customer, @Context HttpHeaders headers){
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.basicAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         try {
             Integer status = customerdao.insert(customer);
             Logback.logger.info("{}.{}|Try: Inserted!", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName());
@@ -96,7 +109,10 @@ public class CustomerRst {
     @Path("/update")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response update(Customer customer){
+    public Response update(Customer customer, @Context HttpHeaders headers){
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.basicAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         try {
             Customer updatedCustomer = customerdao.findById(customer.getCustomerNumber());
             updatedCustomer.setCustomerName(customer.getCustomerName());
@@ -119,7 +135,9 @@ public class CustomerRst {
     @DELETE
     @Path("/{customerNumber}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response delete(@PathParam("customerNumber") Integer customerNumber){
+    public Response delete(@PathParam("customerNumber") Integer customerNumber, @Context HttpHeaders headers){
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.basicAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
         try {
             Integer status = customerdao.delete(customerNumber);
             Logback.logger.info("{}.{}|Try:Deleted", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName());
