+++ Order/src/main/java/com/helman/Webservice/OfficeRst.java
@@ -17,6 +17,8 @@ import com.helman.Entity.Office;
 import com.helman.General.Logback;
 
 import javax.ws.rs.*;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.List;
@@ -24,41 +26,49 @@ import java.util.List;
 @Path("/office")
 public class OfficeRst {
     Officedao officedao = new Officedao();
+    Security sec = new Security();
 
     //http:localhost:8080/order/rest/office/all
     @GET
     @Path("/all")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response findall(){
-        List<Office> officeList = officedao.findAll();
+    public Response findall(@Context HttpHeaders headers){
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ","");
+        if (!sec.basicAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         try{
+            List<Office> officeList = officedao.findAll();
+            Office office = new Office();
             FilterProvider filters = new SimpleFilterProvider().addFilter("Officefilter",
-                    SimpleBeanPropertyFilter.filterOutAllExcept("officeCode", "city", "phone", "country"));
+                    SimpleBeanPropertyFilter.filterOutAllExcept(office.getfilters()));
             String officeJson = (new ObjectMapper()).writer(filters).withDefaultPrettyPrinter().writeValueAsString(officeList);
             Logback.logger.info("{}.{}|Try:Send all records to RESTful", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName());
             return Response.status(Response.Status.OK).entity(officeJson).build();
-        }catch (JsonProcessingException e){
+        }catch (Exception e){
             Logback.logger.error("{}.{}|Exception:{}", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName(), e.getMessage());
             e.printStackTrace();
-            return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+            return Response.status(Response.Status.UNAUTHORIZED).entity(e.getMessage()).build();
         }
     }
     //http:localhost:8080/order/rest/office/find/10
     @GET
     @Path("/find/{officeCode}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response findbyid(@PathParam("officeCode") String ofcode){
-        Office office = officedao.findById(ofcode);
+    public Response findbyid(@PathParam("officeCode") String ofcode, @Context HttpHeaders headers){
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.tokenAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         try{
+            Office office = officedao.findById(ofcode);
             FilterProvider filters = new SimpleFilterProvider().addFilter("Officefilter",
-                    SimpleBeanPropertyFilter.filterOutAllExcept("officeCode", "city", "phone", "addressLine1", "addressLine2", "state", "country", "postalCode"));
+                    SimpleBeanPropertyFilter.filterOutAllExcept(office.getfilters()));
             String officeJson = (new ObjectMapper()).writer(filters).withDefaultPrettyPrinter().writeValueAsString(office);
             Logback.logger.info("{}.{}|Try:send a record to RESTful", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName());
             return Response.status(Response.Status.OK).entity(officeJson).build();
-        }catch (JsonProcessingException e){
+        }catch (Exception e){
             Logback.logger.error("{}.{}|Exception:{}", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName(),e.getMessage());
             e.printStackTrace();
-            return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+            return Response.status(Response.Status.UNAUTHORIZED).entity(e.getMessage()).build();
         }
     }
     /*http://localhost:8080/order/rest/office/insert
@@ -77,7 +87,10 @@ public class OfficeRst {
     @Path("/insert")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response insert(Office office){
+    public Response insert(Office office, @Context HttpHeaders headers){
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.tokenAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or passsword is wrong").build();
+
        try{
            String status = officedao.insert(office);
            Logback.logger.info("{}.{}|Try:Inserted!", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName());
@@ -93,19 +106,12 @@ public class OfficeRst {
     @Path("/update")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response update(Office office){
+    public Response update(Office office, @Context HttpHeaders headers){
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.tokenAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         try{
-            Office updatedOffice = officedao.findById(office.getOfficeCode());
-            updatedOffice.setOfficeCode(office.getOfficeCode());
-            updatedOffice.setCity(office.getCity());
-            updatedOffice.setPhone(office.getPhone());
-            updatedOffice.setAddressLine1(office.getAddressLine1());
-            updatedOffice.setAddressLine2(office.getAddressLine2());
-            updatedOffice.setState(office.getState());
-            updatedOffice.setCountry(office.getCountry());
-            updatedOffice.setPostalCode(office.getPostalCode());
-            updatedOffice.setTerritory(office.getTerritory());
-            String status = officedao.update(updatedOffice);
+            String status = officedao.update(officedao.findById(office.getOfficeCode()));
             return Response.status(Response.Status.OK).entity(status).build();
         }catch (Exception e){
                 Logback.logger.error("{}.{}|Exception:{}", this.getClass().getSimpleName(), Thread.currentThread().getStackTrace()[1].getMethodName(), e.getMessage());
@@ -117,7 +123,10 @@ public class OfficeRst {
     @DELETE
     @Path("/{officeCode}")
     @PathParam(MediaType.APPLICATION_JSON)
-    public Response delete(@PathParam("officeCode") String ofcode){
+    public Response delete(@PathParam("officeCode") String ofcode, @Context HttpHeaders headers){
+        String encodUsrPwd = headers.getRequestHeader("Authorization").get(0).replaceFirst("Basic ", "");
+        if (!sec.tokenAuthCheck(encodUsrPwd)) return Response.status(Response.Status.UNAUTHORIZED).entity("User or password is wrong").build();
+
         try{
             Integer status = officedao.delete(officedao.findById(ofcode));
             return Response.status(Response.Status.OK).entity(status).build();
